Refactor comms tests
- make test related class names more descriptive

- group Listener to Sender tests

- add comment to StateSender on initial state behaviour
