+++ packages/comms/lib/src/bloc/state_sender.dart
@@ -1,6 +1,9 @@
 part of '../../comms.dart';
 
 /// Sends emitted [State]s to all [Listener]s of type [State].
+///
+/// The `initialState` passed in constructor will not get sent. If needed, it
+/// has to be sent manually.
 mixin StateSender<State> on BlocBase<State> {
   @override
   @mustCallSuper
