+++ packages/comms/test/comms_test.dart
@@ -1,11 +1,11 @@
 import 'package:comms/comms.dart';
 import 'package:test/test.dart';
 
-import 'listeners/product_count.dart';
-import 'listeners/product_count_cubit.dart';
+import 'listeners/product_count_listener.dart';
+import 'listeners/product_count_listener_cubit.dart';
 import 'messages/product_count_changed.dart';
-import 'senders/basket.dart';
-import 'senders/basket_cubit.dart';
+import 'senders/basket_cubit_sender.dart';
+import 'senders/basket_sender.dart';
 
 int numberOfProductCountMessageSink() =>
     MessageSinkRegister().getSinksOfType<ProductCountChangedMessage>().length;
@@ -13,147 +13,149 @@ int numberOfProductCountMessageSink() =>
 void main() {
   setUp(() => MessageSinkRegister().clear());
 
-  test(
-    'ProductCount message sink is added to register after it calls listen in constructor',
-    () {
-      final productCount = ProductCount();
-      expect(numberOfProductCountMessageSink(), 1);
-      productCount.dispose();
-    },
-  );
+  group('Listener - Sender', () {
+    test(
+      'ProductCountListener message sink is added to register after it calls listen in constructor',
+      () {
+        final productCount = ProductCountListener();
+        expect(numberOfProductCountMessageSink(), 1);
+        productCount.dispose();
+      },
+    );
 
-  test(
-    'ProductCount message sink is removed from register after it calls cancel in dispose',
-    () {
-      final productCount = ProductCount();
-      expect(numberOfProductCountMessageSink(), 1);
-      productCount.dispose();
-      expect(numberOfProductCountMessageSink(), 0);
-    },
-  );
+    test(
+      'ProductCountListener message sink is removed from register after it calls cancel in dispose',
+      () {
+        final productCount = ProductCountListener();
+        expect(numberOfProductCountMessageSink(), 1);
+        productCount.dispose();
+        expect(numberOfProductCountMessageSink(), 0);
+      },
+    );
 
-  test(
-    'ProductCount value is consistent with number of elements in Basket',
-    () async {
-      final basket = Basket();
-      final productCount = ProductCount();
+    test(
+      'ProductCountListener value is consistent with number of elements in BasketSender',
+      () async {
+        final basket = BasketSender();
+        final productCount = ProductCountListener();
 
-      basket
-        ..add('T-shirt')
-        ..add('Socks');
-      await Future<void>.delayed(Duration.zero);
+        basket
+          ..add('T-shirt')
+          ..add('Socks');
+        await Future<void>.delayed(Duration.zero);
 
-      expect(productCount.value, basket.products.length);
+        expect(productCount.value, basket.products.length);
 
-      basket.removeLast();
-      await Future<void>.delayed(Duration.zero);
+        basket.removeLast();
+        await Future<void>.delayed(Duration.zero);
 
-      expect(productCount.value, basket.products.length);
+        expect(productCount.value, basket.products.length);
 
-      productCount.dispose();
-    },
-  );
+        productCount.dispose();
+      },
+    );
 
-  test(
-    'ProductCount value correctly sets initial state using buffered message',
-    () async {
-      final basket = Basket()..add('Jeans');
-      final productCount = ProductCount();
+    test(
+      'ProductCountListener value correctly sets initial state using buffered message',
+      () async {
+        final basket = BasketSender()..add('Jeans');
+        final productCount = ProductCountListener();
 
-      basket
-        ..add('T-shirt')
-        ..add('Socks');
-      await Future<void>.delayed(Duration.zero);
+        basket
+          ..add('T-shirt')
+          ..add('Socks');
+        await Future<void>.delayed(Duration.zero);
 
-      expect(productCount.value, basket.products.length);
+        expect(productCount.value, basket.products.length);
 
-      basket.removeLast();
-      await Future<void>.delayed(Duration.zero);
+        basket.removeLast();
+        await Future<void>.delayed(Duration.zero);
 
-      expect(productCount.value, basket.products.length);
+        expect(productCount.value, basket.products.length);
 
-      productCount.dispose();
-    },
-  );
+        productCount.dispose();
+      },
+    );
 
-  test(
-    'ProductCountIncrementedListener listens only for ProductCountIncremented',
-    () async {
-      final basket = IncrementingBasket();
-      final productCount = ProductCountIncrementedListener();
+    test(
+      'ProductCountIncrementedListener listens only for ProductCountIncremented',
+      () async {
+        final basket = IncrementingBasket();
+        final productCount = ProductCountIncrementedListener();
 
-      basket
-        ..add('Jeans')
-        ..add('T-shirt');
+        basket
+          ..add('Jeans')
+          ..add('T-shirt');
 
-      await Future<void>.delayed(Duration.zero);
+        await Future<void>.delayed(Duration.zero);
 
-      expect(productCount.value, 2);
+        expect(productCount.value, 2);
 
-      getSend<ProductCountDecremented>()(ProductCountDecremented());
-      await Future<void>.delayed(Duration.zero);
+        getSend<ProductCountDecremented>()(ProductCountDecremented());
+        await Future<void>.delayed(Duration.zero);
 
-      expect(productCount.value, 2);
+        expect(productCount.value, 2);
 
-      getSend<ProductCountChangedMessage>()(ProductCountIncremented());
-      await Future<void>.delayed(Duration.zero);
+        getSend<ProductCountChangedMessage>()(ProductCountIncremented());
+        await Future<void>.delayed(Duration.zero);
 
-      expect(productCount.value, 3);
+        expect(productCount.value, 3);
 
-      getSend<ProductCountChangedMessage>()(ProductCountChangedMessage());
-      await Future<void>.delayed(Duration.zero);
+        getSend<ProductCountChangedMessage>()(ProductCountChangedMessage());
+        await Future<void>.delayed(Duration.zero);
 
-      expect(productCount.value, 3);
+        expect(productCount.value, 3);
 
-      productCount.dispose();
-    },
-  );
+        productCount.dispose();
+      },
+    );
 
-  test(
-    'ProductCountIncrementedListener receives buffered message of exact same type',
-    () async {
-      final basket = IncrementingBasket()..add('Product');
-      final productCount = ProductCountIncrementedListener();
+    test(
+      'ProductCountIncrementedListener receives buffered message of exact same type',
+      () async {
+        final basket = IncrementingBasket()..add('Product');
+        final productCount = ProductCountIncrementedListener();
 
-      basket
-        ..add('Jeans')
-        ..add('T-shirt');
+        basket
+          ..add('Jeans')
+          ..add('T-shirt');
 
-      await Future<void>.delayed(Duration.zero);
+        await Future<void>.delayed(Duration.zero);
 
-      expect(productCount.value, 3);
+        expect(productCount.value, 3);
 
-      productCount.dispose();
-    },
-  );
+        productCount.dispose();
+      },
+    );
 
-  test(
-    'ProductCount receives buffered message of sub type',
-    () async {
-      getSend<ProductCountIncremented>()(ProductCountIncremented());
-      final productCount = ProductCount();
+    test(
+      'ProductCountListener receives buffered message of sub type',
+      () async {
+        getSend<ProductCountIncremented>()(ProductCountIncremented());
+        final productCount = ProductCountListener();
 
-      await Future<void>.delayed(Duration.zero);
+        await Future<void>.delayed(Duration.zero);
 
-      expect(productCount.value, 1);
+        expect(productCount.value, 1);
 
-      productCount.dispose();
-    },
-  );
+        productCount.dispose();
+      },
+    );
 
-  test(
-    'ProductCountIncrementedListener does not receive buffered message of super type',
-    () async {
-      getSend<ProductCountChangedMessage>()(ProductCountChangedMessage());
-      final productCount = ProductCountIncrementedListener();
+    test(
+      'ProductCountIncrementedListener does not receive buffered message of super type',
+      () async {
+        getSend<ProductCountChangedMessage>()(ProductCountChangedMessage());
+        final productCount = ProductCountIncrementedListener();
 
-      await Future<void>.delayed(Duration.zero);
+        await Future<void>.delayed(Duration.zero);
 
-      expect(productCount.value, 0);
+        expect(productCount.value, 0);
 
-      productCount.dispose();
-    },
-  );
+        productCount.dispose();
+      },
+    );
+  });
 
   group('ListenerCubit - Sender:', () {
     test(
@@ -176,9 +178,9 @@ void main() {
     );
 
     test(
-      'ProductCountListenerCubit state is consistent with number of elements in BasketCubit state',
+      'ProductCountListenerCubit state is consistent with number of elements in BasketCubitSender state',
       () async {
-        final basketCubit = BasketCubit();
+        final basketCubit = BasketCubitSender();
         final productCountCubit = ProductCountListenerCubit();
 
         basketCubit
@@ -201,7 +203,7 @@ void main() {
     test(
       'ProductCountListenerCubit correctly sets initial state using buffered message',
       () async {
-        final basketCubit = BasketCubit()..add('Jeans');
+        final basketCubit = BasketCubitSender()..add('Jeans');
         final productCountCubit = ProductCountListenerCubit();
 
         basketCubit
