+++ packages/comms/test/senders/basket_cubit_sender.dart
@@ -2,9 +2,9 @@ import 'package:bloc/bloc.dart';
 import 'package:comms/comms.dart';
 import '../messages/product_count_changed.dart';
 
-class BasketCubit extends Cubit<List<String>>
+class BasketCubitSender extends Cubit<List<String>>
     with Sender<ProductCountChangedMessage> {
-  BasketCubit() : super([]);
+  BasketCubitSender() : super([]);
 
   void add(String product) {
     emit([...state, product]);
